LG_ES_v2.0/requirements.txt
# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4

# Pipeline (same as v1.0_2602)
pandas>=1.5.0
pymysql>=1.0.0
cryptography>=41.0.0